Fixed integer division issues in py3

+++ bluesky/sim/qtgl/screenio.py
@@ -54,11 +54,11 @@ class ScreenIO(QObject):
         self.slow_timer.timeout.connect(self.send_siminfo)
         self.slow_timer.timeout.connect(self.send_aman_data)
         self.slow_timer.timeout.connect(self.send_route_data)
-        self.slow_timer.start(1000 / self.siminfo_rate)
+        self.slow_timer.start(int(1000 / self.siminfo_rate))
 
         self.fast_timer = Timer()
         self.fast_timer.timeout.connect(self.send_aircraft_data)
-        self.fast_timer.start(1000 / self.acupdate_rate)
+        self.fast_timer.start(int(1000 / self.acupdate_rate))
 
     def update(self):
         if bs.sim.state == bs.sim.op:
